pyproject.toml
[project]
name = "haggl"
version = "0.2.0"
description = "Haggl - AI-powered B2B procurement with x402 payment authorization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
# Real CDP integration (optional for demos)
cdp = [
    "cdp-sdk>=0.1.0",
]
# Full production stack
full = [
    "cdp-sdk>=0.1.0",
    "anthropic>=0.40.0",
    "pymongo>=4.6.0",
    "playwright>=1.40.0",
    "stripe>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/calling_agent", "src/x402", "src/payment_agent"]
